Moved IIssueSpanCalculator to IssueViz.
No code changes.
