+++ src/Integration.Vsix/SonarLintTagger/IssueToIssueMarkerConverter.cs
@@ -22,6 +22,7 @@ using System;
 using Microsoft.VisualStudio.Text;
 using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Integration.Suppression;
+using SonarLint.VisualStudio.IssueVisualization.Editor;
 
 namespace SonarLint.VisualStudio.Integration.Vsix
 {
